Refactor + cleanup sidebar and drawpanel buttons
