--- app/src/components/sidebar/DownloadBtn.js
@@ -1,14 +1,0 @@
-import React from 'react';
-
-export default function DownloadBtn({ localDownload }) {
-  const downloadImg = (e) => {
-    e.preventDefault();
-    // localDownload();
-  };
-
-  return (
-    <button onClick={downloadImg}>
-      Download scribble
-    </button>
-  );
-}
