+++ app/src/routes/Main.js
@@ -1,13 +1,11 @@
 import React from 'react';
 
 import DrawPanel from '../components/DrawPanel';
-import Sidebar from '../components/Sidebar';
 
 export default function Main() {
   return (
     <div className="main-page">
       <DrawPanel />
-      <Sidebar />
     </div>
   );
 }
